Move save_logs to tools.lib.helpers
