+++ selfdrive/debug/run_process_on_route.py
@@ -2,10 +2,10 @@
 
 import argparse
 
-from selfdrive.test.compare_logs import save_log
 from selfdrive.test.process_replay.process_replay import CONFIGS, replay_process
 from tools.lib.logreader import MultiLogIterator
 from tools.lib.route import Route
+from tools.lib.helpers import save_log
 
 if __name__ == "__main__":
   parser = argparse.ArgumentParser(description="Run process on route and create new logs",
