+++ selfdrive/test/process_replay/test_processes.py
@@ -9,11 +9,12 @@ from typing import Any, DefaultDict, Dict
 
 from selfdrive.car.car_helpers import interface_names
 from selfdrive.test.openpilotci import get_url, upload_file
-from selfdrive.test.compare_logs import compare_logs, save_log
+from selfdrive.test.process_replay.compare_logs import compare_logs
 from selfdrive.test.process_replay.process_replay import CONFIGS, PROC_REPLAY_DIR, FAKEDATA, check_openpilot_enabled, replay_process
 from system.version import get_commit
 from tools.lib.filereader import FileReader
 from tools.lib.logreader import LogReader
+from tools.lib.helpers import save_log
 
 source_segments = [
   ("BODY", "937ccb7243511b65|2022-05-24--16-03-09--1"),        # COMMA.BODY
