+++ tools/plotjuggler/juggle.py
@@ -11,10 +11,10 @@ import requests
 import argparse
 
 from common.basedir import BASEDIR
-from selfdrive.test.compare_logs import save_log
 from selfdrive.test.openpilotci import get_url
 from tools.lib.logreader import LogReader
 from tools.lib.route import Route, SegmentName
+from tools.lib.helpers import save_log
 from urllib.parse import urlparse, parse_qs
 
 juggle_dir = os.path.dirname(os.path.realpath(__file__))
